requirements.txt
numpy==1.21.5
pyqt 5.9.2
pyqtgraph==0.11.0
yaml==0.2.5
scipy==1.7.3
matplotlib==3.5.1